Feat/#57 일기 및 일기장 작성자인지 검증하는 코드 구현

+++ src/main/java/com/pyro/yolog/domain/diary/service/DiaryService.java
@@ -1,5 +1,6 @@
 package com.pyro.yolog.domain.diary.service;
 
+import com.pyro.yolog.domain.auth.service.AuthService;
 import com.pyro.yolog.domain.diary.dto.request.DiaryContentRequest;
 import com.pyro.yolog.domain.diary.dto.request.DiaryDateRequest;
 import com.pyro.yolog.domain.diary.dto.request.MoodRequest;
@@ -8,13 +9,13 @@ import com.pyro.yolog.domain.diary.dto.response.DefaultDiaryResponse;
 import com.pyro.yolog.domain.diary.dto.response.DetailDiaryResponse;
 import com.pyro.yolog.domain.diary.dto.response.PreviewDiaryResponse;
 import com.pyro.yolog.domain.diary.entity.Diary;
-import com.pyro.yolog.domain.diary.entity.Mood;
-import com.pyro.yolog.domain.diary.entity.Weather;
+import com.pyro.yolog.domain.diary.exception.DiaryNotFoundException;
 import com.pyro.yolog.domain.diary.mapper.DiaryMapper;
 import com.pyro.yolog.domain.diary.repository.DiaryRepository;
+import com.pyro.yolog.domain.member.exception.OwnerNotEqualException;
 import com.pyro.yolog.domain.trip.entity.Trip;
+import com.pyro.yolog.domain.trip.exception.TripNotFoundException;
 import com.pyro.yolog.domain.trip.repository.TripRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -29,24 +30,30 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class DiaryService {
     private static final String DAY = "Day";
+
+    private final AuthService authService;
     private final DiaryRepository diaryRepository;
     private final TripRepository tripRepository;
     private final DiaryMapper diaryMapper;
 
     public DetailDiaryResponse getDiary(Long id) {
-        return new DetailDiaryResponse((diaryRepository.findById(id)
-                .orElseThrow(EntityNotFoundException::new)));
+        Diary diary = diaryRepository.findById(id).orElseThrow(DiaryNotFoundException::new);
+        checkDiaryOwner(diary);
+        return new DetailDiaryResponse(diary);
     }
 
 
     public List<PreviewDiaryResponse> getDiaries(Long tripId, String dayName) {
+        checkTripOwner(tripRepository.findById(tripId)
+                .orElseThrow(TripNotFoundException::new));
         List<Diary> diaries = diaryRepository.findAllByTripIdAndDayName(tripId, dayName);
-        return diaries.stream().map(PreviewDiaryResponse::new).collect(Collectors.toList());
+        return diaries.stream().map(PreviewDiaryResponse::new)
+                .collect(Collectors.toList());
     }
 
     @Transactional
     public DefaultDiaryResponse createDefaultDiary(final Long tripId, final DiaryDateRequest request) {
-        Trip trip = tripRepository.findById(tripId).orElseThrow(EntityNotFoundException::new);
+        Trip trip = tripRepository.findById(tripId).orElseThrow(TripNotFoundException::new);
         final String dayName = DAY + (ChronoUnit.DAYS.between(trip.getStartDate(), request.getDate()) + 1);
         return diaryMapper.toDefaultFormatResponse(
                 diaryRepository.save(diaryMapper.toEntity(trip, dayName, request.getDate())));
@@ -54,30 +61,35 @@ public class DiaryService {
 
     @Transactional
     public void updateDiaryTitleAndContent(Long id, DiaryContentRequest request) {
-        final Diary diary = diaryRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        final Diary diary = diaryRepository.findById(id).orElseThrow(DiaryNotFoundException::new);
+        checkDiaryOwner(diary);
         diary.updateTitleAndContent(request);
     }
 
     @Transactional
     public void deleteDiary(Long id) {
+        checkDiaryOwner(diaryRepository.findById(id).orElseThrow(DiaryNotFoundException::new));
         diaryRepository.deleteById(id);
     }
 
     @Transactional
     public void updateWeather(Long id, WeatherRequest request) {
-        diaryRepository.findById(id).orElseThrow(EntityNotFoundException::new)
-                .updateWeather(request.getWeather());
+        Diary diary = diaryRepository.findById(id).orElseThrow(DiaryNotFoundException::new);
+        checkDiaryOwner(diary);
+        diary.updateWeather(request.getWeather());
     }
 
     @Transactional
     public void updateMood(Long id, MoodRequest request) {
-        diaryRepository.findById(id).orElseThrow(EntityNotFoundException::new)
-                .updateMood(request.getMood());
+        Diary diary = diaryRepository.findById(id).orElseThrow(DiaryNotFoundException::new);
+        checkTripOwner(diary.getTrip());
+        diary.updateMood(request.getMood());
     }
 
 
     @Transactional
     public void deleteOutOfDuration(Trip trip) {
+        checkTripOwner(trip);
         diaryRepository.findById(trip.getId()).ifPresent(diary -> {
             if (diary.getTravelDate().isBefore(trip.getStartDate()) || diary.getTravelDate().isAfter(trip.getFinishDate())) {
                 diaryRepository.deleteById(diary.getId());
@@ -85,4 +97,13 @@ public class DiaryService {
         });
     }
 
+    private void checkDiaryOwner(Diary diary) {
+        checkTripOwner(diary.getTrip());
+    }
+
+    private void checkTripOwner(Trip trip) {
+        if (!trip.getMember().equals(authService.getLoginUser())) {
+            throw new OwnerNotEqualException();
+        }
+    }
 }
